LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/rendering/widgets/shadow-dom-expected.txt



FAIL <video controls=""></video> has a shadow tree with no slots assert_equals: child should be outside the flat tree expected 0 but got 443
FAIL <video></video> has a shadow tree with no slots assert_equals: child should be outside the flat tree expected 0 but got 443
FAIL <audio></audio> has a shadow tree with no slots assert_equals: child should be outside the flat tree expected 0 but got 443
FAIL <audio controls=""></audio> has a shadow tree with no slots assert_equals: child should be outside the flat tree expected 0 but got 444
PASS <select></select> has a shadow tree with slot
PASS <select multiple=""></select> has a shadow tree with slot
PASS <select size="4"></select> has a shadow tree with slot
PASS <details></details> has a shadow tree with slot
PASS <details open=""></details> has a shadow tree with slot
